Update Unidim time series IRT both.stan
This model place actors that answer the same question with both continuous and binary outcome variables on the same dimension. This is used to place the population in cantons and the parties based on their position on direct democratic ballots where they answer the exact same question. It also models the time variance in actor's position as normally distributed around the position at a previous time.